stuffs are almost done for new version
